Feat: 프로필 닉네임/이미지 설정 화면 Base 코드 적용. View-Controller 분리
