+++ meaning-out/Controllers/ProfileNicknameImage/ProfileNicknameViewController.swift
@@ -10,18 +10,9 @@ import UIKit
 import SnapKit
 import TextFieldEffects
 
-class ProfileNicknameViewController: UIViewController {
+class ProfileNicknameViewController: BaseViewController {
     
-    let profileImageView = UIView()
-    let profileImage = UIImageView()
-    
-    let cameraImageView = UIView()
-    let cameraImage = UIImageView()
-    
-    let nicknameField = HoshiTextField()
-    let invalidMessage = UILabel()
-    
-    let doneButton = PointButton(title: Constants.Button.done.rawValue)
+    let nicknameView = ProfileNicknameView()
     
     // 닉네임 유효성 검사 여부
     var isValidate = false
@@ -30,21 +21,21 @@ class ProfileNicknameViewController: UIViewController {
     var isUser = UserDefaultsManager.isUser
     var profileNum = UserDefaultsManager.profile {
         didSet {
-            profileImage.image = Resource.Images.profiles[profileNum]
+            nicknameView.profileImage.image = Resource.Images.profiles[profileNum]
         }
+    }
+    
+    override func loadView() {
+        self.view = nicknameView
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        configureView()
-        configureHierarchy()
-        configureLayout()
         configureUI()
         configureData()
         configureHandler()
     }
-    
+ 
     // 화면 전환 후 다시 돌아왔을 때 필요한 부분 처리
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
@@ -53,88 +44,9 @@ class ProfileNicknameViewController: UIViewController {
         profileNum = UserDefaultsManager.profile
     }
     
-    private func configureView() {
-        view.backgroundColor = Resource.Colors.white
+    override func configureViewController() {
         navigationItem.title = Constants.Title.profile.rawValue
-        
         addImgBarBtn(image: Resource.SystemImages.left, style: .plain, target: self, action: #selector(backBarButtonClicked), type: .left)
-    }
-    
-    private func configureHierarchy() {
-        cameraImageView.addSubview(cameraImage)
-        
-        profileImageView.addSubview(profileImage)
-        profileImageView.addSubview(cameraImageView)
-        
-        let subviews: [UIView] = [profileImageView, nicknameField, invalidMessage, doneButton]
-        subviews.forEach { subview in
-            view.addSubview(subview)
-        }
-    }
-    
-    private func configureLayout() {
-        profileImageView.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide).inset(16)
-            $0.size.equalTo(100)
-            $0.centerX.equalTo(view.safeAreaLayoutGuide)
-        }
-        
-        profileImage.snp.makeConstraints {
-            $0.top.bottom.equalTo(profileImageView)
-            $0.size.equalTo(profileImageView)
-            $0.centerX.equalTo(profileImageView)
-        }
-        
-        cameraImageView.snp.makeConstraints {
-            $0.trailing.equalTo(profileImage)
-            $0.bottom.equalTo(profileImage).inset(8)
-            $0.size.equalTo(24)
-        }
-        
-        cameraImage.snp.makeConstraints {
-            $0.center.equalTo(cameraImageView)
-            $0.size.equalTo(15)
-        }
-        
-        nicknameField.snp.makeConstraints {
-            $0.top.equalTo(profileImageView.snp.bottom).offset(24)
-            $0.horizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(24)
-            $0.height.equalTo(50)
-        }
-        
-        invalidMessage.snp.makeConstraints {
-            $0.top.equalTo(nicknameField.snp.bottom).offset(8)
-            $0.horizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(24)
-            $0.height.equalTo(30)
-        }
-        
-        doneButton.snp.makeConstraints {
-            $0.top.equalTo(invalidMessage.snp.bottom).offset(16)
-            $0.horizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(16)
-            $0.height.equalTo(50)
-        }
-    }
-    
-    private func configureUI() {
-        profileImage.backgroundColor = Resource.Colors.white
-        profileImage.clipsToBounds = true
-        profileImage.layer.cornerRadius = 50  // 추후 상수화 하기
-        profileImage.layer.borderColor = Resource.Colors.primary.cgColor
-        profileImage.layer.borderWidth = CGFloat(Constants.Integer.borderWidth.rawValue)
-        profileImage.contentMode = .scaleAspectFit
-        
-        cameraImageView.backgroundColor = Resource.Colors.primary
-        cameraImageView.layer.cornerRadius = 12  // 추후 상수화 하기
-        
-        cameraImage.image = Resource.SystemImages.camara
-        cameraImage.contentMode = .scaleAspectFit
-        cameraImage.tintColor = Resource.Colors.white
-        
-        nicknameField.setTextFieldUI(Constants.Placeholder.nickname.rawValue)
-        nicknameField.returnKeyType = .done
-        
-        invalidMessage.textColor = Resource.Colors.primary
-        invalidMessage.font = Resource.Fonts.regular13
     }
     
     private func configureData() {
@@ -144,20 +56,19 @@ class ProfileNicknameViewController: UIViewController {
         }
         
         UserDefaultsManager.profile = profileNum
-        profileImage.image = Resource.Images.profiles[profileNum]
+        nicknameView.profileImage.image = Resource.Images.profiles[profileNum]
     }
     
     private func configureHandler() {
         // 프로필 이미지 탭
         let profileTapped = UITapGestureRecognizer(target: self, action: #selector(profileTapped))
-        profileImageView.addGestureRecognizer(profileTapped)
-        profileImageView.isUserInteractionEnabled = true
+        nicknameView.profileImageView.addGestureRecognizer(profileTapped)
+        nicknameView.profileImageView.isUserInteractionEnabled = true
         
         // 닉네임 유효성 검사, 키보드 내리기
-        nicknameField.addTarget(self, action: #selector(validateNickname), for: .editingChanged)
-        nicknameField.addTarget(self, action: #selector(keyboardDismiss), for: .editingDidEndOnExit)
-        
-        doneButton.addTarget(self, action: #selector(doneButtonClicked), for: .touchUpInside)
+        nicknameView.nicknameField.addTarget(self, action: #selector(validateNickname), for: .editingChanged)
+        nicknameView.nicknameField.addTarget(self, action: #selector(keyboardDismiss), for: .editingDidEndOnExit)
+        nicknameView.doneButton.addTarget(self, action: #selector(doneButtonClicked), for: .touchUpInside)
     }
     
     // MARK: @objc 함수
@@ -171,29 +82,29 @@ class ProfileNicknameViewController: UIViewController {
     }
     
     @objc func validateNickname() {
-        guard let nickname = nicknameField.text else { return }
+        guard let nickname = nicknameView.nicknameField.text else { return }
         
         do {
             let result = try getValidationResult(nickname)
             if result {
                 isValidate = true
-                invalidMessage.text = Constants.Validation.Nickname.success.rawValue
+                nicknameView.invalidMessage.text = Constants.Validation.Nickname.success.rawValue
             }
         } catch ValidationError.empty {
             isValidate = false
-            invalidMessage.text = Constants.Validation.Nickname.empty.rawValue
+            nicknameView.invalidMessage.text = Constants.Validation.Nickname.empty.rawValue
         } catch ValidationError.hasSpecialChar{
             isValidate = false
-            invalidMessage.text = Constants.Validation.Nickname.hasSpecialChar.rawValue
+            nicknameView.invalidMessage.text = Constants.Validation.Nickname.hasSpecialChar.rawValue
         } catch ValidationError.hasNumber {
             isValidate = false
-            invalidMessage.text = Constants.Validation.Nickname.hasNumber.rawValue
+            nicknameView.invalidMessage.text = Constants.Validation.Nickname.hasNumber.rawValue
         } catch ValidationError.invalidLength {
             isValidate = false
-            invalidMessage.text = Constants.Validation.Nickname.invalidLength.rawValue
+            nicknameView.invalidMessage.text = Constants.Validation.Nickname.invalidLength.rawValue
         } catch {
             isValidate = false
-            invalidMessage.text = Constants.Validation.Nickname.etc.rawValue
+            nicknameView.invalidMessage.text = Constants.Validation.Nickname.etc.rawValue
         }
     }
     
@@ -212,7 +123,7 @@ class ProfileNicknameViewController: UIViewController {
         }
         
         // 유효성 검사 통과 시 프로필 사진/닉네임 저장
-        guard let nickname = nicknameField.text else { return }
+        guard let nickname = nicknameView.nicknameField.text else { return }
         UserDefaultsManager.nickname = nickname
         UserDefaultsManager.profile = profileNum
         UserDefaultsManager.joinDate = getTodayString(formatType: "yyyy. MM. dd")
